refactor transform code into own module
